backend/pyproject.toml
[project]
name = "scoring-bowlards-backend"
version = "1.0.0"
description = "ボーリングスコア記録・管理アプリケーションのFastAPIバックエンド"
authors = [
    {name = "System Engineer", email = "engineer@example.com"}
]
# readme = "README.md"  # Dockerビルド時にファイルが存在しないためコメントアウト
requires-python = ">=3.13"
dependencies = [
    # FastAPI and ASGI server
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    
    # Firebase and Google Cloud
    "firebase-admin==6.2.0",
    "google-cloud-firestore==2.13.1",
    "google-auth==2.23.4",
    
    # Authentication and Security
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "passlib[bcrypt]==1.7.4",
    
    # Data validation and serialization
    "pydantic==2.10.0",
    "pydantic-settings==2.1.0",
    "email-validator==2.1.0",
    
    # HTTP client
    "httpx==0.25.2",
    
    # Logging and monitoring
    "structlog==23.2.0",
    "prometheus-client==0.19.0",
    
    # Environment management
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "pytest-xdist==3.3.1",
    
    # Code quality
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "pre-commit==3.6.0",
    
    # Documentation
    "mkdocs==1.5.3",
    "mkdocs-material==9.4.8",
    
    # Development tools
    "ipython==8.17.2",
    "jupyter==1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "pytest-xdist==3.3.1",
    
    # Code quality
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "pre-commit==3.6.0",
    
    # Documentation
    "mkdocs==1.5.3",
    "mkdocs-material==9.4.8",
    
    # Development tools
    "ipython==8.17.2",
    "jupyter==1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "firebase_admin.*",
    "google.cloud.*",
    "google.oauth2.*",
    "structlog.*",
    "prometheus_client.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
